% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakup_spt.R
\name{breakup_spt}
\alias{breakup_spt}
\title{detect_sleepbout}
\usage{
breakup_spt(
  sleepBinary = c(),
  tsSPT = c(),
  wakeBoutThreshold = 0.3,
  maxLengthWake = 10,
  sleepBoutMin = 180,
  epochSize = 5
)
}
\arguments{
\item{sleepBinary}{Numeric vector with for each epoch in time a 1 for sleep or a 0 for wakefulness}

\item{tsSPT}{Numeric vector with for each epoch in time a 1 for SPT or a 0 for wake window}

\item{wakeBoutThreshold}{Number between 0 and 1 being the allowed ratio of wakefulness per sleep bout}

\item{maxLengthWake}{Maximum duration of a wakefulness bout in minutes (wake shorter than this will be ignored)}

\item{sleepBoutMin}{Minimum duration of a sleep bout}

\item{epochSize}{Epoch size in seconds}
}
\value{
Data.frame bouts, with for each bout start index, end index, duration, state being
sleep or wake, and sleepRatio being the average ratio of sleep across all of the rolling
30 minute windows in a bout
}
\description{
detect_sleepbout
}
