% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_sleepbout.R
\name{detect_sleepbout}
\alias{detect_sleepbout}
\title{detect_sleepbout}
\usage{
detect_sleepbout(
  sleepBinary = c(),
  wakeBoutThreshold = 0.3,
  wakeBoutMin = 10,
  sleepBoutMin = 180,
  epochSize = 5
)
}
\arguments{
\item{sleepBinary}{Numeric vector with for each epoch in time a 1 for sleep or a 0 for wakefulness}

\item{wakeBoutThreshold}{Number between 0 and 1 being the maximum ratio of wakefulness allowed per 30 minutes}

\item{wakeBoutMin}{Maximum duration of of a wakefulness bout}

\item{sleepBoutMin}{Minimum duration of a sleep bout}

\item{epochSize}{Epoch size in seconds}
}
\value{
Data.frame bouts, with for each bout start index, end index, duration, and state being sleep or wake
}
\description{
detect_sleepbout
}
