% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bouts.R
\name{extract_bouts}
\alias{extract_bouts}
\title{extract_bouts}
\usage{
extract_bouts(x)
}
\arguments{
\item{x}{Numeric vector with for each epoch in time a 1 a 0}
}
\value{
Data.frame bouts, with for each bout start index, end index, duration, state being
1 or 2
}
\description{
extract_bouts
}
