% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sleepbout_column.R
\name{add_sleepbout_column}
\alias{add_sleepbout_column}
\title{add_bouts}
\usage{
add_sleepbout_column(
  ts = c(),
  sleepBinary = c(),
  wakeBoutThreshold = 0.3,
  wakeBoutMin = 10,
  sleepBoutMin = 180,
  epochSize = c(),
  maxLengthWake = 10,
  detection.method = 2
)
}
\arguments{
\item{ts}{Time series dataframe produced by GGIR that has column window to indicate wakingup-wakingup window and column class_id, where 0 reflects sleep}

\item{sleepBinary}{See \link{detect_sleepbout}}

\item{wakeBoutThreshold}{See \link{detect_sleepbout}}

\item{wakeBoutMin}{See \link{detect_sleepbout}}

\item{sleepBoutMin}{See \link{detect_sleepbout}}

\item{epochSize}{See \link{detect_sleepbout}}

\item{maxLengthWake}{See \link{breakup_spt}}

\item{detection.method}{1 for bout detection, 2 for SPT break up}
}
\value{
Input data frame expanded with one column to indicate where the sleep bouts occur by a number which reflects the sleep bouw number per window
}
\description{
add_bouts
}
