% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sleepbout_column.R
\name{add_sleepbout_column}
\alias{add_sleepbout_column}
\title{add_bouts}
\usage{
add_sleepbout_column(ts = c(), epochSize = c())
}
\arguments{
\item{ts}{Time series dataframe produced by GGIR that has column window to indicate wakingup-wakingup window and column class_id, where 0 reflects sleep}

\item{epochSize}{epoch size in seconds}
}
\value{
Input data frame expanded with one column to indicate where the sleep bouts occur by a number which reflects the sleep bouw number per window
}
\description{
add_bouts
}
